AppMobike/requirements.txt
api2pdf==0.0.11
arabic-reshaper==2.1.1
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
cryptography==3.2.1
defusedxml==0.6.0
Django==3.1.1
django-allauth==0.43.0
django-credit-cards==0.4.1
django-extensions==3.0.9
django-leaflet==0.27.1
django-sslserver==0.22
django-wkhtmltopdf==3.3.0
future==0.18.2
gcloud==0.18.3
googleapis-common-protos==1.52.0
html5lib==1.1
httplib2==0.18.1
idna==2.10
isort==5.5.4
jws==0.1.3
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauth2client==4.1.2
oauthlib==3.1.0
pdfkit==0.6.1
Pillow==8.0.1
protobuf==3.14.0rc1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.9
PyJWT==1.7.1
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyOpenSSL==19.1.0
PyPDF2==1.26.0
Pyrebase4==4.4.1
python-bidi==0.4.2
python-jwt==2.0.1
python3-openid==3.2.0
pytz==2020.1
reportlab==3.5.55
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.6
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.11
webencodings==0.5.1
wrapt==1.12.1
xhtml2pdf==0.2.5
